# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Era Task API Documentation",
        "description": "API documentation for authentication and admin management.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/segments": {
            "get": {
                "tags": [
                    "Segments"
                ],
                "summary": "List all segments with pagination",
                "operationId": "20dfeed7bba402127ef8e0493879b25f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated segments list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SegmentListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Segments"
                ],
                "summary": "New segment generate",
                "operationId": "9d9e1c7825ab2fa9024da9260a502409",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSegmentSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Segment uğurla yaradıldı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Aktiv istifadəçilər"
                                        },
                                        "filter_json": {
                                            "type": "object",
                                            "example": {
                                                "user": {
                                                    "email_verified": true
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/segments/{id}": {
            "get": {
                "tags": [
                    "Segments"
                ],
                "summary": "Get segment details by ID",
                "description": "Returns a specific segment by its ID",
                "operationId": "a03b59530a493e40e0123e371d5d8db6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Segment ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Segment found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SegmentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Segment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Segment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogoutErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/segments/{id}/preview": {
            "get": {
                "tags": [
                    "Segments"
                ],
                "summary": "Preview a segment",
                "operationId": "770a0762640e94517c6ec13ed37b1c60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Segment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Segment preview",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SegmentPreviewResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Segment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Segment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "LogoutErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "CreateSegmentSchema": {
                "title": "Create Segment",
                "description": "Yeni bir seqment yaratmaq üçün məlumat strukturu",
                "required": [
                    "name",
                    "filter_json"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Aktiv istifadəçilər"
                    },
                    "filter_json": {
                        "type": "object",
                        "example": {
                            "user": {
                                "email_verified": true,
                                "marketing_opt_in": false
                            },
                            "product": {
                                "price_min": 10,
                                "price_max": 100
                            }
                        }
                    }
                },
                "type": "object"
            },
            "SegmentListResponse": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Aktiv istifadəçilər"
                                },
                                "filter_json": {
                                    "type": "object",
                                    "example": {
                                        "email_verified": true,
                                        "marketing_opt_in": true
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "first_page_url": {
                        "type": "string",
                        "example": "/api/v1/segments?page=1"
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "last_page_url": {
                        "type": "string",
                        "example": "/api/v1/segments?page=5"
                    },
                    "next_page_url": {
                        "type": "string",
                        "example": "/api/v1/segments?page=2"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "example": null
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "total": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
            "SegmentPreviewResponse": {
                "properties": {
                    "total_recipients": {
                        "type": "integer",
                        "example": 2841
                    },
                    "sample": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 123
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Aylin M."
                                },
                                "email": {
                                    "type": "string",
                                    "example": "aylin@example.com"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "SegmentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Active Users Segment"
                    },
                    "filter_json": {
                        "type": "object",
                        "example": {
                            "email_verified": true,
                            "marketing_opt_in": true
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-19T12:34:56Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000oauth/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}